fix erc1155 preset/base suite separation
